+++ EstatisticaFatec.Core/Models/VariavelQuantitativa/VariavelQuantitativaEntity.cs
@@ -9,12 +9,12 @@
         /// <summary>
         /// [1] - Dado bruto do input
         /// </summary>
-        public int XI { get; set; }
+        public decimal XI { get; set; }
 
         /// <summary>
         /// [2] - Quantidade de elementos [XI]
         /// </summary>
-        public int FI { get; set; }
+        public decimal FI { get; set; }
 
         /// <summary>
         /// [3] - Porcentagem baseando na [quantidade total de FI] sobre o [FI]
@@ -24,7 +24,7 @@
         /// <summary>
         /// [4] - Números acumulativos sobre [FI]
         /// </summary>
-        public int F { get; set; }
+        public decimal F { get; set; }
 
         /// <summary>
         /// [5] - Porcentagem acumulativa sobre [FE%]
